pyproject.toml
[project]
name = "hapztext-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daphne>=4.2.1",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-eventstream[drf]>=5.3.2",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular[sidecar]>=0.28.0",
    "loguru>=0.7.3",
    "nanoid>=2.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyotp>=2.9.0",
    "redis>=6.4.0",
    "whitenoise>=6.9.0",
]
